Return UserDTO also when asking for Friends
