+++ back/src/main/java/coronhackathon/backend/DTO/UserAndNbChallengeDTO.java
@@ -1,0 +1,24 @@
+package coronhackathon.backend.DTO;
+
+public class UserAndNbChallengeDTO {
+    private long userId;
+    private String username;
+    private long nbChall;
+    public UserAndNbChallengeDTO(long userId, String username, long nbChall){
+        this.userId = userId;
+        this.username = username;
+        this.nbChall = nbChall;
+    }
+
+    public long getNbChall() {
+        return nbChall;
+    }
+
+    public long getUserId() {
+        return userId;
+    }
+
+    public String getUsername() {
+        return username;
+    }
+}
